Backend/requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
Flask>=2.3
google-auth==2.40.3
google-auth-oauthlib==1.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
urllib3==2.5.0
Werkzeug>=3.1.3,<4.0.0
flask-mysql-connector==1.1.0
python-dotenv==1.0.0
flask-cors==3.0.10
flask-jwt-extended==4.4.4
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
sqlalchemy>=2.0,<3.0
bcrypt==4.0.1
pytest==7.4.0
pytest-flask==1.2.0
python-dateutil==2.8.2
flask-bcrypt==1.0.1
flask-mysqldb==0.2.0
mysqlclient==2.1.1
aniso8601==8.0.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy>=3.0.0
pytz==2020.1
six==1.15.0
pytesseract>=0.3.10,<0.4.0
mysql-connector-python
pymysql
cryptography